ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
pillow==10.2.0
pytesseract==0.3.10
joblib==1.3.2
opencv-python-headless==4.9.0.80
scipy==1.11.4
matplotlib==3.8.2
python-dotenv==1.0.0
